+++ lib/Utility/DetailsHelper.php
@@ -7,7 +7,7 @@
  * later. See the COPYING file.
  *
  * @author Pauli Järvinen <pauli.jarvinen@gmail.com>
- * @copyright Pauli Järvinen 2018 - 2020
+ * @copyright Pauli Järvinen 2018 - 2024
  */
 
 namespace OCA\Music\Utility;
@@ -27,11 +27,6 @@ class DetailsHelper {
 		$this->logger = $logger;
 	}
 
-	/**
-	 * @param integer $fileId
-	 * @param Folder $userFolder
-	 * $return array|null
-	 */
 	public function getDetails(int $fileId, Folder $userFolder) : ?array {
 		$file = $userFolder->getById($fileId)[0] ?? null;
 		if ($file !== null) {
@@ -91,15 +86,14 @@ class DetailsHelper {
 	}
 
 	/**
-	 * @param integer $fileId
-	 * @param Folder $userFolder
-	 * $return string|null
+	 * Get lyrics from the file metadata in plain-text format. If there's no unsynchronised lyrics available
+	 * but there is synchronised lyrics, then the plain-text format is converted from the synchronised lyrics.
 	 */
-	public function getLyrics(int $fileId, Folder $userFolder) {
+	public function getLyricsAsPlainText(int $fileId, Folder $userFolder) : ?string {
 		$lyrics = null;
-		$fileNodes = $userFolder->getById($fileId);
-		if (\count($fileNodes) > 0) {
-			$data = $this->extractor->extract($fileNodes[0]);
+		$fileNode = $userFolder->getById($fileId)[0] ?? null;
+		if ($fileNode) {
+			$data = $this->extractor->extract($fileNode);
 			$lyrics = ExtractorGetID3::getFirstOfTags($data, ['unsynchronised_lyric', 'unsynced lyrics', 'unsynced_lyrics', 'unsyncedlyrics']);
 			self::sanitizeString($lyrics);
 
@@ -118,6 +112,43 @@ class DetailsHelper {
 	}
 
 	/**
+	 * Get all lyrics from the file metadata, both synced and unsynced. For both lyrics types, a single instance
+	 * of lyrics contains an array of lyrics lines. In case of synced lyrics, the key of this array is the offset
+	 * in milliseconds.
+	 * @return array of items like ['synced' => boolean, 'lines' => array]
+	 */
+	public function getLyricsAsStructured(int $fileId, Folder $userFolder) : array {
+		$result = [];
+		$fileNode = $userFolder->getById($fileId)[0] ?? null;
+		if ($fileNode) {
+			$data = $this->extractor->extract($fileNode);
+			$lyricsTags = ExtractorGetID3::getTags($data, ['LYRICS', 'lyrics', 'unsynchronised_lyric', 'unsynced lyrics', 'unsynced_lyrics', 'unsyncedlyrics']);
+
+			foreach ($lyricsTags as $tagKey => $tagValue) {
+				self::sanitizeString($tagValue);
+
+				// Never try to parse synced lyrics from the "unsync*" tags. The "lyrics" tag, on the other hand,
+				// may contain either synced or unsynced lyrics and a parse attempt is needed to find out.
+				$mayBeSynced = !Util::startsWith($tagKey, 'unsync');
+				$syncedLyrics = $mayBeSynced ? LyricsParser::parseSyncedLyrics($tagValue) : null;
+
+				if ($syncedLyrics) {
+					$result[] = [
+						'synced' => true,
+						'lines' => $syncedLyrics
+					];
+				} else {
+					$result[] = [
+						'synced' => false,
+						'lines' => \explode("\n", $tagValue)
+					];
+				}
+			}
+		}
+		return $result;
+	}
+
+	/**
 	 * Read lyrics-related tags, and build a result array containing potentially
 	 * both time-synced and unsynced lyrics. If no lyrics tags are found, the result will
 	 * be null. In case the result is non-null, there is always at least the key 'unsynced'
@@ -125,9 +156,6 @@ class DetailsHelper {
 	 * If found and successfully parsed, there will be also another key 'synced', which will
 	 * hold the time-synced lyrics. These are presented as an array of arrays of form
 	 * ['time' => int (ms), 'text' => string].
-	 *
-	 * @param array $tags
-	 * @return array|null
 	 */
 	private static function transformLyrics(array $tags) : ?array {
 		$lyrics = $tags['LYRICS'] ?? $tags['lyrics'] ?? null; // may be synced or unsynced
@@ -184,8 +212,6 @@ class DetailsHelper {
 	/**
 	 * In the 'comments' field from the extractor, the value for each key is a 1-element
 	 * array containing the actual tag value. Remove these intermediate arrays.
-	 * @param array $array
-	 * @return array
 	 */
 	private static function flattenComments(array $array) : array {
 		// key 'text' is an exception, its value is an associative array